// src/api/schema.d.ts
export interface components {
    schemas: {
        ApiResponse: {
            /** Format: int32 */
            code?: number;
            type?: string;
            message?: string;
        };
        Category: {
            /** Format: int64 */
            id?: number;
            name?: string;
        };
        Pet: {
            /** Format: int64 */
            id?: number;
            category?: components["schemas"]["Category"];
            /** @example doggie */
            name: string;
            photoUrls: string[];
            tags?: components["schemas"]["Tag"][];
            /**
             * @description pet status in the store
             * @enum {string}
             */
            status?: "available" | "pending" | "sold";
        };
        Tag: {
            /** Format: int64 */
            id?: number;
            name?: string;
        };
        Order: {
            /** Format: int64 */
            id?: number;
            /** Format: int64 */
            petId?: number;
            /** Format: int32 */
            quantity?: number;
            /** Format: date-time */
            shipDate?: string;
            /**
             * @description Order Status
             * @enum {string}
             */
            status?: "placed" | "approved" | "delivered";
            complete?: boolean;
        };
        User: {
            /** Format: int64 */
            id?: number;
            username?: string;
            firstName?: string;
            lastName?: string;
            email?: string;
            password?: string;
            phone?: string;
            /**
             * Format: int32
             * @description User Status
             */
            userStatus?: number;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: {
        /** @description List of user object */
        UserArray: {
            content: {
                "application/json": components["schemas"]["User"][];
            };
        };
        /** @description Pet object that needs to be added to the store */
        Pet: {
            content: {
                "application/json": components["schemas"]["Pet"];
                "application/xml": components["schemas"]["Pet"];
            };
        };
    };
    headers: never;
    pathItems: never;
}